requirements.txt
streamlit==1.28.1
PyPDF2==3.0.1
python-docx==0.8.11
nltk==3.8.1
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
spacy==3.6.1
textblob==0.17.1
plotly==5.15.0
sqlite3